+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Bug fixes:
 
 Improvements:
 
+  - [compl] Allow named param completion on functions
   - [wr] Infer return type for generators
   - [lsp] Show Phpactor version info in initialize result @lalanikarim
   - [ls] Fixed class completion performance